+++ src/MainWindow.h
@@ -7,22 +7,19 @@
 #include <QtWidgets/QErrorMessage>
 #include <QtWidgets/QProgressDialog>
 #include <QDropEvent>
+#include <osgViewer/Viewer>
 
 #include "ui_MainWindow.h"
-#include "OSGViewerWidget.h"
-#include "dragLabel.h"
-#include "narrative/NarrativeInfoDialog.h"
-#include "dragLabelInput.h"
-#include "labelCanvas.h"
-#include "labelCanvasView.h"
 
 //#include "VSimApp.h"
-
 //extern osgViewer::Viewer* g_viewer;
 
 class VSimApp;
+class OSGViewerWidget;
 class labelCanvas;
 class labelCanvasView;
+class ModelOutliner;
+class TimeSlider;
 
 class MainWindow : public QMainWindow
 {
@@ -38,7 +35,12 @@ public:
 
 	// linking
 	//OSGViewerWidget *getViewerWidget() const { return m_osg_widget;	}
-	osgViewer::Viewer* getViewer() const { return m_osg_widget->getViewer(); };
+	osgViewer::Viewer* getViewer() const;
+
+	MainWindowTopBar *topBar() const;
+	labelCanvasView *canvasView() const;
+	ModelOutliner *outliner() const;
+	TimeSlider *timeSlider() const;
 
 	// event stuff
 	void paintEvent(QPaintEvent* event);
@@ -66,16 +68,24 @@ signals:
 	void sSaveCurrent();
 	void sNew();
 	void sImportModel(const std::string&);
+	void sImportNarratives();
+	void sExportNarratives();
+
+	void sDebugOSG();
+	void sDebugCamera();
 
 	// ui, signal emitters
 public:
-	Ui::MainWindow ui;
+	std::unique_ptr<Ui::MainWindow> ui;
 
 public:
+	QUndoStack *m_undo_stack;
+
 	OSGViewerWidget *m_osg_widget;
 	labelCanvas *m_drag_area;
-	QUndoStack *m_undo_stack;
 	labelCanvasView *m_view;
+	ModelOutliner *m_outliner;
+	TimeSlider *m_time_slider;
 };