+++ src/dragLabel.cpp
@@ -3,9 +3,9 @@
 #include <QtGui/QMouseEvent>
 #include <iostream>
 #include <QtCore/QBasicTimer>
-#include "MainWindow.h"
 #include "dragLabel.h"
 #include "mrichtextedit.h"
+#include "dragLabelInput.h"
 
 //constructor for novel generation
 dragLabel::dragLabel(labelCanvas* parent, std::string style)